[crash-reduce] NFC: Split up `ProcessReproducers.process`
